add native token

+++ helix-contract/contracts/mapping-token/interfaces/IOutboundLane.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.8.0;
 
 interface IOutboundLane {
-    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint256);
+    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint64);
 }
